requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
motor
python-dotenv
email-validator
cloudinary
jinja2
certifi